Separated failure utility functions out from Runner
Moved them to a new file as standalone functions so they're not tied to the `Runner` logic.

Continues #6.
